requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
psycopg2-binary==2.9.10
requests==2.32.0
python-dotenv==1.0.0
